requirements.txt
beautifulsoup4==4.12.2
certifi==2023.11.17
charset-normalizer==3.3.2
exceptiongroup==1.2.0
furl==2.1.3
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
lxml==5.0.1
orderedmultidict==1.0.1
pillow==10.2.0
requests==2.31.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
typing_extensions==4.9.0
urllib3==2.1.0
